Fix: 동적 렌더링 설정 변경
